SV-23 Fixed time-variant tests. Updated integration test path. Move getTask usecase to api dir.
