+++ .env.sample
@@ -1,6 +1,6 @@
 DB_USER=
 DB_PASSWORD=
-API_PATH=http://localhost:3000/dev/api/healthcheck
+API_PATH=http://localhost:3000/dev/api
 UI_PATH=http://localhost:3000
 JWT_SECRET=secret
 ALLOWED_GROUPS=housing-officer-dev,area-housing-manager-dev
