+++ package.json
@@ -6,6 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
+    "serverless:offline": "npm run build && serverless offline",
     "test:unit": "jest --config ./jest.unit.config.js",
     "test:unit:watch": "jest --config ./jest.unit.config.js --watch",
     "test:unit:coverage": "jest --config ./jest.unit.config.js --coverage",
